refactor(photo-curation): rename judge→fake-judge and sync.test→sources.sync.test
judge.ts shadowed the judges/ directory in fuzzy-finders and editors:
both names resolve when typing "judge", but judge.ts is the deterministic
test-double while judges/ holds the real Gemini implementations. Renaming
to fake-judge.ts makes the distinction unambiguous without moving anything.

sync.test.ts tested sync/syncAll/scoreBatch from sources.ts but was named
after a module that does not exist — renamed to sources.sync.test.ts so
the filename names its subject.

Three import specifiers updated (./judge.js → ./fake-judge.js) in
fake-judge.test.ts, sources.test.ts, and sources.sync.test.ts.
No export names changed; no logic changed.
