+++ tools/photo-curation/src/fake-judge.ts
